+++ src/gridstack-engine.ts
@@ -1,4 +1,4 @@
-// gridstack-engine.ts 3.1.5-dev @preserve
+// gridstack-engine.ts 3.2.0 @preserve
 
 /**
  * https://gridstackjs.com/