+++ src/gridstack.ts
@@ -1,4 +1,4 @@
-// gridstack.ts 3.1.5-dev @preserve
+// gridstack.ts 3.2.0 @preserve
 
 /**
  * https://gridstackjs.com/