+++ src/h5/dd-base-impl.ts
@@ -1,4 +1,4 @@
-// dd-base-impl.ts 3.1.5-dev @preserve
+// dd-base-impl.ts 3.2.0 @preserve
 
 /**
  * https://gridstackjs.com/