+++ src/h5/dd-resizable-handle.ts
@@ -1,4 +1,4 @@
-// dd-resizable-handle.ts 3.1.5-dev @preserve
+// dd-resizable-handle.ts 3.2.0 @preserve
 
 /**
  * https://gridstackjs.com/