+++ src/h5/dd-utils.ts
@@ -1,4 +1,4 @@
-// dd-utils.ts 3.1.5-dev @preserve
+// dd-utils.ts 3.2.0 @preserve
 
 /**
  * https://gridstackjs.com/